[fix] Workaround for array values with illegal indices in models
